Add files via upload
I would like to make an observation here and I apologize if this is not the appropriate space to do so. When I upload my Excel file on Google Drive, some sheets, graphs, charts and tables change completely for some reason. I hope this do not happen when I upload them on GitHub. Could you please contact me if this does happen? Please see some examples in the Word document. Many thanks!